#186, #216: adding grant types constants and separting classes inside OpenIdConnectConstants.cs
