+++ src/Microsoft.IdentityModel.Protocols.OpenIdConnect/OpenIdConnectParameterNames.cs
@@ -1,0 +1,74 @@
+﻿//-----------------------------------------------------------------------
+// Copyright (c) Microsoft Open Technologies, Inc.
+// All Rights Reserved
+// Apache License 2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//-----------------------------------------------------------------------
+
+namespace Microsoft.IdentityModel.Protocols.OpenIdConnect
+{
+    /// <summary>
+    /// Parameter names for OpenIdConnect.
+    /// </summary>
+    public static class OpenIdConnectParameterNames
+    {
+#pragma warning disable 1591
+        public const string AccessToken = "access_token";
+        public const string AcrValues = "acr_values";
+        public const string ClaimsLocales = "claims_locales";
+        public const string ClientAssertion = "client_assertion";
+        public const string ClientAssertionType = "client_assertion_type";
+        public const string ClientId = "client_id";
+        public const string ClientSecret = "client_secret";
+        public const string Code = "code";
+        public const string Display = "display";
+        public const string DomainHint = "domain_hint";
+        public const string Error = "error";
+        public const string ErrorDescription = "error_description";
+        public const string ErrorUri = "error_uri";
+        public const string ExpiresIn = "expires_in";
+        public const string GrantType = "grant_type";
+        public const string Iss = "iss";
+        public const string IdToken = "id_token";
+        public const string IdTokenHint = "id_token_hint";
+        public const string IdentityProvider = "identity_provider";
+        public const string LoginHint = "login_hint";
+        public const string MaxAge = "max_age";
+        public const string Nonce = "nonce";
+        public const string Password = "password";
+        public const string PostLogoutRedirectUri = "post_logout_redirect_uri";
+        public const string Prompt = "prompt";
+        public const string RedirectUri = "redirect_uri";
+        public const string RequestUri = "request_uri";
+        public const string Resource = "resource";
+        public const string ResponseMode = "response_mode";
+        public const string ResponseType = "response_type";
+        public const string Scope = "scope";
+        public const string SessionState = "session_state";
+        public const string State = "state";
+        public const string TargetLinkUri = "target_link_uri";
+        public const string Token = "token";
+        public const string TokenType = "token_type";
+        public const string UiLocales = "ui_locales";
+        public const string UserId = "user_id";
+        public const string Username = "username";
+#pragma warning restore 1591
+    }
+}
+
+
+
+
+
+
