+++ src/Microsoft.IdentityModel.Protocols.OpenIdConnect/OpenIdConnectScopes.cs
@@ -1,0 +1,32 @@
+﻿//-----------------------------------------------------------------------
+// Copyright (c) Microsoft Open Technologies, Inc.
+// All Rights Reserved
+// Apache License 2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//-----------------------------------------------------------------------
+
+namespace Microsoft.IdentityModel.Protocols.OpenIdConnect
+{
+    /// <summary>
+    /// Specific scope values that are interesting to OpenID Connect.  See http://openid.net/specs/openid-connect-messages-1_0.html#scopes
+    /// </summary>
+    public static class OpenIdConnectScopes
+    {
+        #pragma warning disable 1591
+        public const string OpenId = "openid";
+        public const string OpenIdProfile = "openid profile";
+        public const string UserImpersonation = "user_impersonation";
+        #pragma warning restore 1591
+    }
+}
